Restructure worker flow. Remove concavs and holes.

+++ assembly/rect.ts
@@ -5,44 +5,44 @@ export default class Rect {
   private _topRight: Point = new Point();
   private _size: Point = new Point();
 
-  constructor(x: f32 = 0, y: f32 = 0, width: f32 = 0, height: f32 = 0) {
+  constructor(x: f64 = 0, y: f64 = 0, width: f64 = 0, height: f64 = 0) {
     this._bottomLeft.update(x, y);
     this._topRight.update(x + width, y + height);
     this._size.update(width, height);
   }
 
-  public get x(): f32 {
+  public get x(): f64 {
     return this._bottomLeft.x;
   }
 
-  public set x(value: f32) {
+  public set x(value: f64) {
     this._bottomLeft.x = value;
     this._topRight.x = this._size.x + value;
   }
 
-  public get y(): f32 {
+  public get y(): f64 {
     return this._bottomLeft.y;
   }
 
-  public set y(value: f32) {
+  public set y(value: f64) {
     this._bottomLeft.y = value;
     this._topRight.y = this._size.y + value;
   }
 
-  public get width(): f32 {
+  public get width(): f64 {
     return this._size.x;
   }
 
-  public set width(value: f32) {
+  public set width(value: f64) {
     this._size.x = value;
     this._topRight.x = this._size.x + value;
   }
 
-  public get height(): f32 {
+  public get height(): f64 {
     return this._size.y;
   }
 
-  public set height(value: f32) {
+  public set height(value: f64) {
     this._size.y = value;
     this._topRight.y = this._bottomLeft.y + value;
   }
