+++ assembly/util.ts
@@ -4,8 +4,8 @@ export const TOLEARANCE: f64 = Math.pow(10, -9); // Floating point error is like
 export const POLYGON_CONFIG_SIZE: u16 = 14;
 
 export function almostEqual(
-  a: f32,
-  b: f32,
+  a: f64,
+  b: f64,
   tolerance: f64 = TOLEARANCE
 ): boolean {
   return Math.abs(a - b) < tolerance;
